adding error handler
adding error handler from json input so that we know what error input do we have when testing

+++ service/category_service_implementation.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 	"database/sql"
+	"golang-restfulapi-exercise/exception"
 	"golang-restfulapi-exercise/helper"
 	"golang-restfulapi-exercise/model/domain"
 	"golang-restfulapi-exercise/model/web"
@@ -25,18 +26,20 @@ func NewCategoryService(categoryRepository repository.CategoryRepository, db *sq
 	}
 }
 
-func (service *CategoryServiceImplem) FindByIdGet(ctx context.Context, idKategori int) web.CategoryResponse {
+func (service *CategoryServiceImplem) FindById(ctx context.Context, idKategori int) web.CategoryResponse {
 	tx, error := service.DB.Begin()
 	helper.PanicIfError(error)
 	defer helper.CommitorRollback(tx)
 
 	categoryid, error := service.CategoryRepository.FindById(ctx, tx, idKategori)
-	helper.PanicIfError(error)
+	if error != nil {
+		panic(exception.NewNotFoundError(error.Error()))
+	}
 
 	return helper.ToCategoryResponse(categoryid)
 }
 
-func (service *CategoryServiceImplem) FindAllGet(ctx context.Context) []web.CategoryResponse {
+func (service *CategoryServiceImplem) FindAll(ctx context.Context) []web.CategoryResponse {
 	tx, error := service.DB.Begin()
 	helper.PanicIfError(error)
 	defer helper.CommitorRollback(tx)
@@ -46,7 +49,7 @@ func (service *CategoryServiceImplem) FindAllGet(ctx context.Context) []web.Cate
 	return helper.ToCategoryResponses(categories)
 }
 
-func (service *CategoryServiceImplem) CreatePost(ctx context.Context, request web.CategoryCreateRequest) web.CategoryResponse {
+func (service *CategoryServiceImplem) Create(ctx context.Context, request web.CategoryCreateRequest) web.CategoryResponse {
 	error := service.Validate.Struct(request)
 	helper.PanicIfError(error)
 
@@ -64,7 +67,7 @@ func (service *CategoryServiceImplem) CreatePost(ctx context.Context, request we
 
 }
 
-func (service *CategoryServiceImplem) UpdatePut(ctx context.Context, request web.CategoryUpdateRequest) web.CategoryResponse {
+func (service *CategoryServiceImplem) Update(ctx context.Context, request web.CategoryUpdateRequest) web.CategoryResponse {
 	error := service.Validate.Struct(request)
 	helper.PanicIfError(error)
 
@@ -73,7 +76,9 @@ func (service *CategoryServiceImplem) UpdatePut(ctx context.Context, request web
 	defer helper.CommitorRollback(tx)
 
 	category, error := service.CategoryRepository.FindById(ctx, tx, request.Id)
-	helper.PanicIfError(error)
+	if error != nil {
+		panic(exception.NewNotFoundError(error.Error()))
+	}
 
 	category.Namakategori = request.Namakategori
 
@@ -82,13 +87,15 @@ func (service *CategoryServiceImplem) UpdatePut(ctx context.Context, request web
 	return helper.ToCategoryResponse(category)
 }
 
-func (service *CategoryServiceImplem) DeleteDelete(ctx context.Context, idKategori int) {
+func (service *CategoryServiceImplem) Delete(ctx context.Context, idKategori int) {
 	tx, error := service.DB.Begin()
 	helper.PanicIfError(error)
 	defer helper.CommitorRollback(tx)
 
 	category, error := service.CategoryRepository.FindById(ctx, tx, idKategori)
-	helper.PanicIfError(error)
+	if error != nil {
+		panic(exception.NewNotFoundError(error.Error()))
+	}
 
 	service.CategoryRepository.Delete(ctx, tx, category)
 }
